Update dependencies (not polonius-engine and proc-macro2), and add very simplistic testing scripts to detect output changes.
The scripst are very basic and not documented, but they are helpfull to check if something breaks when updating dependencies.
The two dependencies that are nod updated cause troubles with the currenly used rustc version rustc 1.28.0-nightly. Maybe it is easyer to update polonius-engine only after switching to a new compiler version. And the fixed version for proc-macro2 can hopeuffly be remuved after switching to a newer compiler version.