First steps of adaption for upgraded compiler version. Does not at all work for now.
The module main seems to be in a more or less working state, despite the part that interfaces with dump_borrowck_info. Howeer, maybe it needs more adaptions, e.g. due to the evolvement of oter parts.
In addition, the module driver_utils was disabled, as it is no longer needed as main now connects to rustc_dirver directly.