Fix one more error. It looks like some return type (or argument type?) changed and threfore an aditional transformation is needed.
Figuring this out was quite cumbersome, and took me some time, as I did not find the right clues in the error messgae. (Maybe they were there, but I did not get them)
However, it looks like quite some of the following errors that are reportet can be fixed in a similar way.