Add method to find start/entry points of a set of points based on a given cfg, and refactor the newly written code.
The refactoring introduced some object-oriented ideas and thereby allows to store and share some data in a struct. Therefore it is no longer needed to pass around a huge bunch of arguments, which hopefully increases the readability. (And is also good for the future, as the number of needed information that should be passed around will probably even increase.)