Update searching for the path that leads to an error. Use termination criterion that was discussed this morning.
I.e. check if starting points of the region are intersecting with the starting points of the loan that is involved in causing the error.
This was not jet tested thoroughly at all, but in principle it seems to work out. (Though, there might still be bugs and some small issues remain. E.g. the "very first" region, being R2 for both example3.rs and find_path_in_graph_ex0, is never included in the path that is emitted right now.) Also, this is not jet really complete in the sense that the found path is only printed as debug output for now. (If debug level is set appropriately.)