New termination criterion for finding the error path, (again) now based on the input fact that (directly) maps regions to loans.
This was not jet tested that thoroughly, but short testing with few examples gave rather promising outputs. Note, that outputs still only consist of debug prints. (Only appear when the log level is set appropriately.)
Also, it must be pointed out that the current code war primarily written for testing, and it therefore does not only contain a lot of debug prints, but also still runs an old version and contains some unused methods and commented code. Once the development continues (after some testing), quite some clean-up will be needed. Also, it might be possible to optimize the performance at some points. (But there are more important issues right now.)