Start adding code that does print the path to the error that can now be found as a graph. (Draft, not jet fully working.)
The finding of the path was tested to quite some extent as of now, but evaluating it without seeing it in a graphical way is rather hard.
Instead, the printing part is more of a first draft. It does apparently some work, but some parts are also still missing and some investigations are needed to fill them in. Also, this was not really tested at all. (It does not crash for at least one example.)
Also, some code refactoring was applied. Right now, the error_path is (partially) printed instead of expl_outlives, but this can now be switched easily.