Refine support for printing error path as graph.
This seems to now work fine. The graphs are printed to nll-facts/<function name>/error_graph.dot, and the (old) expl_outlives graph is still printed to outlive_graph.dot.
This was also tested with all examples that were available and working as of now. Looking at the results leads me to the conclusion that the printing seems to work fine. All information that is printed in the expl_outlives_graph and is on the path that was found before is also printed to the error graph. (However, there is still information, e.g. about source lines, missing that was already missing before in the (expl_)outlives_graph.)