Cleanup ErrorPathFinder: Remove or comment out computations and debug outputs that became obsolete.
Quite some things that were computed and/or printed in the ErrorPathFinder are leftovers from past try-outs and obsolete
ways of finding the path. These were not removed or commented out. Also all methods that were not (or no longer) used
were comment out.
In addition, quite some documentation was added.