Implement walking of methods, and fix issues with multiple error points and when no region can be found for an error point.
As of now, methods (functions) that were part of an  block were not handled by the tool, but simply ignored. So if error occurred in these methods, no explaining output war created. This should be fixed by this commit.
Also, there were some issues when there were multiple error points reported by polonius, and it turned out that (for some error points, and in the case of having two errors in the input program) the inference of the region that causes an error can fail. This was also addressed, and the tool does no longer panic in such cases.
Also, two examples (actually adoptions of example9) for these cases were added.