Fix issue regarding nested functions.
This commit (effectively one line in dump_borrowck_info.rs) fixes the issue that functions that were defined and
declared inside of another function (e.g. inside of the main function, as it is the case for some examples that are part
of the error index and other documentation, probably for the sake of easy integration of the rust playground.) were not
handled by the tool.
This seems to work fine, no regressions seemed to occur (as far as tested) and the issue seems to be solved.
Also, the tests that were not usable due to this limitation (and therefore were disabled temporarily, file names ending
with ) were re-enable again. (by renaming the files)