Start fixing linking of regions to source lines.
For now, only a small update that unifies the handling for both regions that are part of an edge in the printed graph. However, this does not cause any improvements (or even changes) in the outputted result.
In addition, a comment now informs about issues in the parsing of the mir for getting information to link regions to locals.
Also a small improvement of the build system (Make file) that was applied in the meeting yesterday is part of this commit.