Try to implement new way to match regions to locals. Does not at all work jet.
Does not work or even really introduce any changes to the resulting graphs. However, it does produce quite some debug output. This output does not jet contain much helpful information, but it helps to show some issues that I have right now, and that hinder me from getting the information that we want to get. So this could help to show what would be needed to make it work.
Also, some new comments in the source etc. give some hints on things that I tried to fix the these issues (without success tough), and provides some extra information/documentation.