Meeting with Federico: Implement better parsing of MIR dump.
Since getting an appropriate version of MIR seems not that easy, we now sicked to keep reading in the needed information from a dump. This is now done with some new RegExes we came up with this morning.
In addition, this was roughly embedded into the existing code, and some old trials were commented out.
However, please note that this was not jet tested, esp. since it is not even really used as of now.