Change print_outlive_error_graph(...) method to use the newly computed region_to_local_map.
This seems to work fine. Also, some other small improvements were applied.
The major change was tested with all examples. It never degraded the results, and sometimes lead to some improvements. I.e. for some examples, more regions were linked to a local (and thereby to a source location) successfully.
Still, this does not fix everything. As expected from looking at the MIR dump, this does still leave some regions unlinked. It is not clear if the parsing of the dump could be improved in ways that lead to significantly more information for linking these regions. As next step, we will probably try to link them by using either the borrow_regions or the requires relation and the existing code to link from points to source lines. (This might be sufficient to complete these results in a sufficient way.)