Add struct to store an enriched error graph/path, and use it.
This also includes adding a method that creates such a struct and initializes it using existing methods or new methods that now offer a concrete, simple functionality that was before part of some other method. Also, switch the method print_outlive_error_graph(...) to use this new struct (passed as argument) and the information it contains to print the graph instead of getting all information right before/while printing.