Implement more graph improvements. More exactly, these remove all regions that are not linked to local with name from the error path, while always preserving the first and the last region of the path.
The functionality is implemented as a function operating on the EnrichedErrorGraph. The resulting graph is now printed to a file named error_graph_improved.dot, next to the old error_graph.dot file that is also printed for comparison, testing and evaluation.
These changes were also tested and shorty evaluated on all examples. The code is executed without any notable problems, and it also seems to improve the resulting graph in the sense that it makes it simpler for quite some examples. Also, it seems that it does not really remove information that was really helpful in the version that was not further optimized.