First stepts to write the VScode plug-in: Get current file and (try to) run extract-error on it. (Does not really work)
The tool (Rust Life/compiler_mod/extract-error) must be located at ~/.rust-life/extract-error to be found by the plug-in. Then it should be run, but it seems that something does not work jet. (At least, I did not find the tool output) It might be that one of the issues it also that the tool always ends with a non-zero return code, but it is also quite likely that there is an issue with running it at all, since the output is not at any place where I would expect it. (So it is probably not created at all.) Unfortunately, I can also not see the stdout/stderr of the tool, it is not printed to the log as expected, unfortunately.