Fix running the tool from the extension. Does in principle work now. (But is not perfect)
There were tow issues, both were trivial to fix, but kind of annoying to spot. So there are tow learnings. First, around running processes from a node.js (VScode extension), ~ is never expanded to the current user's home dir, this must be done manually. Second, I do not exactly want to know how node.js does pass arguments when running an external tool. (Apparently in ways that cause malfunctions when using string concatenation instead of passing each space-separed element as one argument.)
So the tool is run now, but it does not jet really provide all needed outputs. Also, the location of the outputs might not be optimal right now. So it does now work in principle, but there is more work to do.
Btw. it is expected that the tool does terminate with a non-zero exit code for now, but node.js seems not to care about this to much anyway, and we do not use the result for now.