Add support to dump EnrichedErrorGraph as JSON to rust-life tool (compiler mod)
This is done using serialization with serde(_json), and seems to work quite smooth. Most of the code for the serialization is derived automatically this way. However, some small changes to the internals (mainly to the EnrichedErrorGraph structure and it's clients) were needed in order to enable this simple JSON dumping. This changes allow to exclude some information that is not really needed from the dump and thereby easily get around the non-serializeability of structures from MIR and to cope with a limitation of serde when it comes to serialize maps that use tuples as keys. (Serde apparently fails to serialize such maps right now, or I have miss-configured it in some way.)