Add support to show the error visualization. (Very basic layout, not styled at all)
In principle, the displaying seems to work (for simple, standard cases), but it was not extensively tested. Also, pleas do not look for code quality in the extension sources, there isn't any. There are only about 100 lines of TypeScript (that does hacky things that look like plain JavaScript), a lot of hardcoded, inline HTML, some inline-inline CSS and even a bit of JavaScript that is inlined into the HTML that is inlided as string literal.
So this code is really more like a proof of concept, if it should become production ready it will definitely need at least some improvements. (Right now, it violates several guidelines for writing VScode extensions and for especially for using it's WebView functionality.)
Also remember that this extension expects the extract-error executable to be located at ~/.rust-life/extract-error to work.
Oh, and would the author of this commit mind to remove this ugly debug print from this one loop? Or even better, add sensible logging facilities and allow for disabling all debug prints.