Start adding support for highlighting lines on a click on the graph.
The passing over of the information from the WebView to the extension (using rather simplistic message passing) does already seem to work fine. Also, the information about the lines can be read from the enrichedErrorGraph JSON dump successfully for some cases. However, for some cases it misses in the enrichedErrorGraph, and must be added in the rust-life/compiler_mod/extract-error program first. Also, the information about the lines is for now only printed to the log console, the actual highlighting is not jet implemented. (And there might be some more details to consider around it, e.g. when to undo the highlighting.)