Acquire information about line numbers for locals, and display it. (Bot in the dot output and in the IDE extension)
Getting this information was done before anyway, but now it is also stored (to EnrichedErrorGraph). Then, this info is used when generating the dot graph.
It is also used in the IDE extension (as it is part of the JSON output) to display this information. Furthermore, the information is also used in the extension to correctly highlight lines now.
The display and highlighting in the IDE extension was now tested for all examples, and seems to work. (Expect the limitation described in the previous commit, and issues that were not noted jet.)