Enable onClickHandler to refined the editor that they should use, and restore the highlighting when an editor window gets visible again.
This is implemented rather simply, but seems to work fine for simple cases, though it was not that thoroughly tested.
It is known not to work when the file is closed, and then reopened, but this seems to be a fine limitation (for now). But at least it does it's clean-up when the webView panel is closed.
Also, there definitely is quite some room for improvements. (e.g. remove the useless mapping from regions to line, and instead just call the methods from the HTML with the line numbers directly, i.e. just add this bit of information to the HTML.)