Improve line highlighting logic, removes rather useless levels of indirection.
Now, the line that shall be highlighted for each region/constraint is computed ("statically") when creating the graph (HTML) for the WebView, instead for computing this information every time when an onClick event occurs. This is simpler and includes less levels of indirection. (All in all, it definitely reduces the number of needed lines of code.) It might also be more efficient, esp. if the graph is getting bigger. (The handling of onClick highlighting definitely includes less computation now.)