Restructure most of the extension, esp. of the visualization.
This does not incorporate any new functionality. However, the command for starting the visualization as Graph was renamed.
Also, most of the visualization functionality was organized into classes (that even use inheritance) that are now in the file errorVisualization.ts. This restructuring and reorganization should (at least slightly) improve the structure of the code, and maybe also slightly increases the readability. However, mostly it is intended to allow reusing some functionality when introducing a new visualization, i.e. one that does no longer show the error as a graph. This shall be done as next step, and will reuse some methods by inheriting from the Visualization class, likewise to what is done for now by the visualization as graph.
However, maybe one wants to do some more (smaller) restructuring when actually implementing this.