First version that supports text-based error explanation. (Very basic)
The textual error visualization is now available as command "Rust Life: Explain with Text". It is very basic, and completely lacks any formatting or styling, so it does not look pretty at all. (Actually, the text seems to be at least slightly too small to be well-readable) Also, it was not really tested (with more then one example) as of now, and there is probably some evaluation needed to apply some further improvements about the actual elements/parts that are reported. (This could possible be combined with coming up with some more realistic examples for evaluating the tool.)
Also, the textual visualization does not jet include the error message from the compiler, this is left as an open TODO.
Furthermore, this commit does apply some small fixes/changes.