Restructure code to prepare for getting the compiler error message, and fix an issue.
This restructuring does mainly consist of moving around some code, this should allow to implement access to the compiler error message (emitted on stderr by extract-error) in a simple way, and does also give the code a slightly better structure and allow for a little more code reuse.
Furthermore, this commit should fix an issue in the textual error representation, more exactly in it's acquisition of information about locales. It does now also work if the local definitions include a `mut` keyword.