Enable displaying the compiler error in the textual explanation. (Very basic)
This is very basic, but seems to work in principle. Actually, it does simply display the stderr from Rust Life (compiler_mud/extract-error) in the WebView. Implementing this was indeed quite easy and not extremely hacky due to the restructuring that was done before.
This was not really tested for now. (Only with two examples) More testing will be done while evaluation this new way of displaying information. (I.e. evaluating the textual representation.)
Also, there definitely are quite a lot of possibilities for improvements, both related to this change and to the entire extension.