Some (small) improvements for the textual error representation.
The compiler error message is now shown with correct indentation. The texts were optimized to be clearer and simpler, and the subtitle does now mention the variable that is still borrowing the initial variable.
Also, the (rather useless) lines/points that stated things like "y borrows y" are no longer included in the explanation, since they are of no value.
In addition, there should be no more warnings included in the included compiler error messages, since warnings are now disabled when runing rust life. (and hence also for the internal rustc, who’s stderr is printed as error message.)
All of these changes were discussed and proposed in the meeting this morning.