Fix an issue in the rust-life-assistant
This change should resolve crashes of the IDE extension in the case when Rust Life does not create a JSON file as expected.
However, this actually only fixes part of the error, since it does not allow to detect when no new file was created, but there still is an old one from a previous execution.