Fix an issue in the Rust Life Assistant (VScode extension) with "let ref"
This fixes the issue when there is a "ref" keyword after the let.
However, there is another issue, actually the affected line does not even always need a "let" as first token, as the line that is parsed must not necessarily also be the line where the variable is defined, but it might also be a line where the variable is assigned. See e.g. example_ifelse1.rs.
This issue will be fixed in the next commit, and in a following commit some examples that can be used to better test this fix will be added.