Fix the issue (try to) while finding variable names in the Rust Life Assistant (Used for textual representation only, for now.)
This fix seems to work fine, but it was not extensively tested (for now, only with about five examples that seem to cover all relevant cases) since it will definitely be tested anyway by the ongoing evaluation.
We seem to cover more cases now, but it might also introduce new issue with other (corner) cases. Also, it is definitely possible that there are even more cases that we still do not handle correctly.