Create a (proforma) cargo project in the collected_code folder.
This is primary intended to prevent complaints of the Rust (rls) extension for VScode when opening this folder, e.g. when testing the Rust Life Assistant VScode extension.
Maybe it does also allow to use the Rust (rls) extension for support when creating new examples, but this was not jet tested successfully.
Also note that the file src/main.rs (added in this commit) is only there to ensure that the Cargo project does work fine. (It might be possible to circumvent this, but this was not further investigated for now.)