Add some new examples that are based on the ones taken from the error index.
Most of these examples were simply the ones from the error index, that have been extended. In most cases, this means that some more levels of indirection (with the involved variables/borrows/etc.) were added in order to increase the complexity of the errors in the examples, and hence enable the tool to actually be able to analyse an error and hence actually be somehow useful.
Some of the examples are also useful to test the changes/fixes in the Rust Life Assistant (VScode Extension) that were committed before.
Some more information about the changes might also be available in the comments in the new examples.