Add some examples with loops.
The examples seem to be fine, but for some of them Rust Life (Assistant) fails to analyse them. There seems to be some bug in Rust Life, guiding from the log output it does not even succeed in finding the error in the examples. If time permits, this issue could be further analysed, but it might be rather deep in the internals of Rust Life. Otherwise, it should be reported as an example that the tool cannot handle correctly.
In addition, this commit fixes an error in the text in the UI of Rust Life Assistant, more exactly in an error message that is displayed to the user.