Add some more examples, both some using closures, an a large one. (Some reveal errors in Rust Life)
Some of these examples work just fine, and example3_long3.rs is intended to be well suited for showing off the functionality of Rust Life.
However, when it is changed in some ways it will reveal erroneous behaviour of Rust Life. (See files ending in _err) More exactly, the textual representation will contain some uninvolved lines that are not directly of relevance. (But are related to some of the relevant variables.)