Remove unused (often old) graph outputs.
This reduces the number of graphs (and hence files) that will be created for each run.
For now, the coresponding lines only were commentd, most of them (all that are not reusable for debugging) will be removed in a further commit to really reduce the code size. Maybe this also introduces some more unused functions, that will also (eventually) be cleaned up in a next commit.