Remove (comment) a lot of old code that is no longer needed.
Some of it was used for some debug prints, that were removed as well, since the same information is printed also from another source.
This commit also removes (comments) a lot of the logic (and hence code) from the first error explanation graph computation that was commited by Dominik, so this commit somehow also manifests the decision to drop this computation. (Also as an intermediate step.)
The code was only commented for now, it will be deleted in a next commit. (To get a clearer history of these removals.)